feat(client): database shadcn migration
